reuse integer test macros for bnum_integer
